refactor(portcullis): extract kernel test module to kernel_tests.rs
Move 1634-line inline test module from kernel.rs to separate file.
kernel.rs: 3858 → 2225 lines (1633 line reduction, well below 3900 ceiling).

All 67 kernel unit tests + 14 red team tests pass identically.
